refactor: rename plugin
