+++ djangocms_tacc_header_logo/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class TaccsiteHeaderLogoConfig(AppConfig):
+    name = 'djangocms_tacc_header_logo'
+    verbose_name = 'Header logo'
