+++ djangocms_tacc_header_logo/constants.py
