+++ taccsite_cms/templatetags/header_tags.py
@@ -1,7 +1,7 @@
 from django import template
 from django.utils.safestring import mark_safe
 
-from taccsite_cms.contrib.taccsite_header_logo.render import render as render_header
+from djangocms_tacc_header_logo.render import render as render_header
 
 register = template.Library()
